backend/requirements.txt
django~=5.0.2
djangorestframework~=3.14.0
drf-yasg~=1.21.4
django-cors-headers~=4.3.1